Move CompositionPipeline out of the pipeline module.
