+++ components/compositing/pipeline.rs
@@ -2,6 +2,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
+use CompositionPipeline;
 use CompositorProxy;
 use compositor_thread;
 use compositor_thread::Msg as CompositorMsg;
@@ -58,15 +59,6 @@ pub struct Pipeline {
     pub running_animations: bool,
     pub children: Vec<FrameId>,
     pub is_private: bool,
-}
-
-/// The subset of the pipeline that is needed for layer composition.
-#[derive(Clone)]
-pub struct CompositionPipeline {
-    pub id: PipelineId,
-    pub script_chan: IpcSender<ConstellationControlMsg>,
-    pub layout_chan: LayoutControlChan,
-    pub chrome_to_paint_chan: Sender<ChromeToPaintMsg>,
 }
 
 /// Initial setup data needed to construct a pipeline.
